+++ app/merchant/vouchers/components/VoucherCard.tsx
@@ -5,24 +5,11 @@ import { Button } from "@/components/ui/button"
 import { Voucher } from "@/app/vouchers/types/voucher"
 
 interface Props {
-  onCardClick?: (voucher: Voucher) => void
-  onActionClick?: (voucher: Voucher) => void
-  actionLabel?: string
+  voucher: Voucher
 }
 
-export default function VoucherCard({ voucher, onCardClick}: Props) {
+export default function VoucherCard({ voucher}: Props) {
   const router = useRouter()
-
-  const handleCardClick = () => {
-    if (onCardClick) {
-      onCardClick(voucher)
-    } else {
-      router.push(`/merchants/vouchers/details/${voucher.id}`)
-    }
-  }
-
-  const remainingCount = voucher.remainingCount;
-  const fillPercent = Math.round((100 * remainingCount) / voucher.totalCount);
 
 
   return (
@@ -30,7 +17,6 @@ export default function VoucherCard({ voucher, onCardClick}: Props) {
       className="bg-white rounded-xl shadow-sm overflow-hidden cursor-pointer"
       whileHover={{ y: -2 }}
       transition={{ duration: 0.2 }}
-      onClick={handleCardClick}
     >
       <div className="relative h-40">
         <Image
@@ -39,15 +25,6 @@ export default function VoucherCard({ voucher, onCardClick}: Props) {
           fill
           className="object-cover"
         />
-        <div className="absolute top-2 right-2 flex items-center space-x-2 bg-black/70 text-white text-xs px-3 py-1 rounded-full">
-          <span>남은 수량: {remainingCount}개</span>
-          <div className="w-16 bg-gray-300 rounded-full h-2">
-            <div
-              className="h-2 rounded-full bg-[#FFB020]"
-              style={{ width: `${fillPercent}%` }}
-            />
-          </div>
-        </div>
       </div>
 
       <div className="p-4">
@@ -73,13 +50,6 @@ export default function VoucherCard({ voucher, onCardClick}: Props) {
                 </p>
               )}
             </div>
-            <Button
-              size="sm"
-              className="rounded-lg bg-[#FFB020] hover:bg-[#FF9500] text-white"
-              disabled={remainingCount <= 0}
-            >
-              {remainingCount}
-            </Button>
           </div>
         </div>
       </div>
