+++ app/merchant/vouchers/components/VoucherList.tsx
@@ -1,4 +1,4 @@
-import VoucherCard from "@/app/vouchers/components/VoucherCard"
+import VoucherCard from "@/app/merchant/vouchers/components/VoucherCard"
 import { getVouchers } from "@/app/merchant/vouchers/api/voucher"
 import { useEffect, useState } from "react"
 import { useSearchParams } from "next/navigation"
